requirements.txt
amqp==2.6.1
asgiref==3.5.0
billiard==3.6.4.0
celery==4.4.7
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.13
django-phone-verify==0.1.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
kombu==4.6.11
MarkupSafe==2.1.1
packaging==21.3
phonenumbers==8.12.46
phonenumberslite==8.12.46
PyJWT==2.3.0
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
twilio==7.8.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
vine==1.3.0
zipp==3.8.0
